Add files via upload
This group project was completed in accordance with the IE-2042 ( Database Management Systems for Security) for the first semester of the second year of the Cyber Security Specialization program. 

My 2nd year 1st semester project is focused on creating an Online Airline Reservation System. This is a group project that will involve designing and implementing a database to store and manage flight and booking information. The system will also include features such as booking, payment processing, and customer management. The project will also focus on implementing security measures to protect sensitive data and prevent unauthorized access. The goal of the project is to create a reliable and secure online reservation system that simplifies the process of booking flights for customers.

1.Document any assumptions made.
2.Develop the ERD and logical model
3.Normalize the logical model to 3NF.
4.Implement the logical model in MS SQL server and enter suitable sample data
5.Identify all necessary constraints and enforce them on the tables
6.Develop the required views, functions, stored procedures, triggers, and indexes as specified below.
7.Identify 2 suitable triggers that can be applied on the database and explain and implement them
8.Identify the possible users of this database and create 2 views for them.
9.Vulnerabilities focusing on techniques and impact.mitigate the selected vulnerabilities and suggest countermeasures to overcome from the selected vulnerabilities